# Optimize available quantity loader in 3.1