I want to merge this change because copy changes from #7802 to master.
I optimizing `AvailableQuantityByProductVariantIdCountryCodeAndChannelSlugLoader`.

To observe progress of this optimization I measure execution time of `AvailableQuantityByProductVariantIdCountryCodeAndChannelSlugLoader.batch_load_quantities_by_country` 


  | New solution(s) | Old solution(s)
-- | -- | --
1 | 0,057897 | 1,262272
2 | 0,057024 | 1,222058
3 | 0,056215 | 1,224122
4 | 0,056883 | 1,222713
5 | 0,056946 | 1,22418
6 | 0,056331 | 1,21168
7 | 0,056319 | 1,204111
8 | 0,056214 | 1,229411
9 | 0,056546 | 1,200475
10 | 0,057739 | 1,213455
AVG | 0,0568114 | 1,2214477
p95 | 0,0578259 | 1,24748455

# Impact

* [ ] New migrations
* [ ] New/Updated API fields or mutations
* [ ] Deprecated API fields or mutations
* [ ] Removed API types, fields, or mutations
* [ ] Documentation needs to be updated

# Pull Request Checklist

<!-- Please keep this section. It will make the maintainer's life easier. -->

* [ ] Privileged queries and mutations are guarded by proper permission checks
* [ ] Database queries are optimized and the number of queries is constant
* [ ] Database migration files are up to date
* [ ] The changes are tested
* [ ] GraphQL schema and type definitions are up to date
* [ ] Changes are mentioned in the changelog