@patrys 2021-08-19
Task linked: [SALEOR-4177 Slow queries reported by Lush](https://app.clickup.com/t/2549495/SALEOR-4177)